backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv